Remove eatmydata to avoid littering the logs with error messages
At the moment the logs are littered with this error message:

    ERROR: ld.so: object 'libeatmydata.so' from LD_PRELOAD cannot be preloaded (cannot open shared object file): ignored.

This line appears ~100 times in the logs...

This is due to autopkgtest-build-lxc, which tries and fails to load
eatmydata.  Apparently, it's apparmor that prevents preloading
libraries. Apparmor is installed as it's a Recommends of lxc.

eatmydata is installed in all CI images, cf.
https://salsa.debian.org/salsa-ci-team/pipeline/-/blob/e7fd6d7df58b4f554c3c20b81ce63d856fcdcefa/images/files/base/initialize

I don't think we *need* eatmydata here, so let's remove it and remove
this error message from the logs.

Some references:
* https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=963508
* https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1010437